+++ src/lib/commandclass/MultiChannelCC.ts
@@ -10,7 +10,7 @@ import { getEnumMemberName, validatePayload } from "../util/misc";
 import { num2hex } from "../util/strings";
 import { encodeBitMask, Maybe, parseBitMask } from "../values/Primitive";
 import { CCAPI } from "./API";
-import { API, CCCommand, CCCommandOptions, ccKeyValuePair, ccValue, CommandClass, commandClass, CommandClassDeserializationOptions, CommandClassOptions, DynamicCCResponse, expectedCCResponse, gotDeserializationOptions, implementedVersion } from "./CommandClass";
+import { API, CCCommand, CCCommandOptions, ccKeyValuePair, CCResponsePredicate, ccValue, CommandClass, commandClass, CommandClassDeserializationOptions, CommandClassOptions, DynamicCCResponse, expectedCCResponse, gotDeserializationOptions, implementedVersion } from "./CommandClass";
 import { CommandClasses } from "./CommandClasses";
 
 export enum MultiChannelCommand {
@@ -834,21 +834,30 @@ interface MultiChannelCCCommandEncapsulationOptions extends CCCommandOptions {
 	destination: MultiChannelCCDestination;
 }
 
-const getResponseForCommandEncapsulation: DynamicCCResponse = (
-	sent: MultiChannelCCCommandEncapsulation,
+const testResponseForCommandEncapsulation: CCResponsePredicate<MultiChannelCCCommandEncapsulation> = (
+	sent,
+	received,
+	isPositiveTransmitReport,
 ) => {
 	// SDS13783: A receiving node MAY respond to a Multi Channel encapsulated command if the Destination
 	// End Point field specifies a single End Point. In that case, the response MUST be Multi Channel
 	// encapsulated.
 	// A receiving node MUST NOT respond to a Multi Channel encapsulated command if the
 	// Destination End Point field specifies multiple End Points via bit mask addressing.
-	return typeof sent.destination === "number"
-		? MultiChannelCCCommandEncapsulation
-		: undefined;
+	if (typeof sent.destination === "number") {
+		return received instanceof MultiChannelCCCommandEncapsulation &&
+			sent.destination === received.endpointIndex
+			? "checkEncapsulated"
+			: isPositiveTransmitReport
+			? "checkEncapsulated"
+			: "unexpected";
+	} else {
+		return isPositiveTransmitReport ? "final" : "unexpected";
+	}
 };
 
 @CCCommand(MultiChannelCommand.CommandEncapsulation)
-@expectedCCResponse(getResponseForCommandEncapsulation)
+@expectedCCResponse(testResponseForCommandEncapsulation)
 export class MultiChannelCCCommandEncapsulation extends MultiChannelCC {
 	public constructor(
 		driver: Driver,