+++ src/lib/commandclass/WakeUpCC.ts
@@ -4,12 +4,20 @@ import log from "../log";
 import { MessagePriority } from "../message/Constants";
 import type { ZWaveNode } from "../node/Node";
 import { NodeStatus } from "../node/Types";
+import type { ValueID } from "../node/ValueDB";
 import { validatePayload } from "../util/misc";
 import { ValueMetadata } from "../values/Metadata";
 import type { Maybe } from "../values/Primitive";
 import { CCAPI, SetValueImplementation, SET_VALUE, throwUnsupportedProperty, throwWrongValueType } from "./API";
 import { API, CCCommand, CCCommandOptions, ccValue, ccValueMetadata, CommandClass, commandClass, CommandClassDeserializationOptions, expectedCCResponse, gotDeserializationOptions, implementedVersion } from "./CommandClass";
 import { CommandClasses } from "./CommandClasses";
+
+export function getWakeUpIntervalValueId(): ValueID {
+	return {
+		commandClass: CommandClasses["Wake Up"],
+		property: "wakeUpInterval",
+	};
+}
 
 export enum WakeUpCommand {
 	IntervalSet = 0x04,